refactor: clean up imports and enhance component structure in growing unit and location update forms
This commit refactors the import statements in the growing unit detail page and location update form components for improved organization and consistency. It also updates the sidebar template to use a more specific import path for the SidebarProvider. These changes enhance code readability and maintainability.
